requirements.txt
absl-py==0.11.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
dataclasses==0.7
google-auth==1.26.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
h5py==3.1.0
idna==2.10
importlib-metadata==3.4.0
Markdown==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
ordered-set==4.0.2
Pillow==8.1.0
pip==20.3.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
setuptools==52.0.0.post20210125
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchmeta==1.6.1
torchvision==0.8.2
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wheel==0.36.2
zipp==3.4.0
